fix(service): JWT 인증 받은 사용자에 한해 게시글 삭제가 가능하도록 수정

+++ backendProject/src/main/java/org/example/backendproject/board/service/BoardService.java
@@ -76,10 +76,17 @@ public class BoardService {
 
     /** 게시글 삭제 **/
     @Transactional
-    public void deleteBoard(Long boardId) {
-        if (!boardRepository.existsById(boardId)) {
-            throw new IllegalArgumentException("게시글 없음: " + boardId);
+    public void deleteBoard(Long userId, Long boardId) {
+        Board board = boardRepository.findById(boardId)
+            .orElseThrow(() -> new IllegalArgumentException("사용자 정보가 없습니다"));
+
+        if (!board.getUser().getId().equals(userId)) {
+            throw new IllegalArgumentException("삭제 권한이 없습니다.");
         }
+
+        if (!boardRepository.existsById(boardId))
+            throw new IllegalArgumentException("게시글 없음: " + boardId);
+
         boardRepository.deleteById(boardId);
     }
